Change type of bytes from int to size_t

+++ idg-lib/src/common/Types.h
@@ -94,7 +94,7 @@ namespace idg {
     T* allocate_memory(size_t n) {
         void *ptr = nullptr;
         if (n > 0) {
-            int bytes = n * sizeof(T);
+            size_t bytes = n * sizeof(T);
             bytes = (((bytes - 1) / ALIGNMENT) * ALIGNMENT) + ALIGNMENT;
             if (posix_memalign(&ptr, ALIGNMENT, bytes) != 0) {
                 std::cerr << "Could not allocate " << bytes << " bytes" << std::endl;
